+++ src/config.php
@@ -87,7 +87,6 @@ return [
         //    ],
         //    'email' => '',
         //    'apiKey' => '',
-        //    'warmCacheDelay' => '5',
         //],
 
         // The purger type classes to add to the plugin’s default purger types.