+++ src/drivers/warmers/DummyWarmer.php
@@ -28,21 +28,21 @@ class DummyWarmer extends BaseCacheWarmer
     /**
      * @inheritdoc
      */
-    public function warmUris(array $siteUris, callable $setProgressHandler = null, int $delay = null, bool $queue = true)
+    public function warmUris(array $siteUris, callable $setProgressHandler = null, bool $queue = true)
     {
     }
 
     /**
      * @inheritdoc
      */
-    public function warmSite(int $siteId, callable $setProgressHandler = null, int $delay = null, bool $queue = true)
+    public function warmSite(int $siteId, callable $setProgressHandler = null, bool $queue = true)
     {
     }
 
     /**
      * @inheritdoc
      */
-    public function warmAll(callable $setProgressHandler = null, int $delay = null, bool $queue = true)
+    public function warmAll(callable $setProgressHandler = null, bool $queue = true)
     {
     }
 }